Move internal/language/proto to language/proto
