+++ cmd/gazelle/BUILD.bazel
@@ -18,12 +18,12 @@ go_library(
         "//config:go_default_library",
         "//internal/flag:go_default_library",
         "//internal/language/go:go_default_library",
-        "//internal/language/proto:go_default_library",
         "//internal/merger:go_default_library",
         "//internal/version:go_default_library",
         "//internal/walk:go_default_library",
         "//label:go_default_library",
         "//language:go_default_library",
+        "//language/proto:go_default_library",
         "//repos:go_default_library",
         "//resolve:go_default_library",
         "//rule:go_default_library",
