+++ internal/merger/BUILD.bazel
@@ -18,8 +18,8 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//internal/language/go:go_default_library",
-        "//internal/language/proto:go_default_library",
         "//language:go_default_library",
+        "//language/proto:go_default_library",
         "//rule:go_default_library",
     ],
 )
