+++ internal/merger/merger_test.go
@@ -19,8 +19,8 @@ import (
 	"testing"
 
 	"github.com/bazelbuild/bazel-gazelle/internal/language/go"
-	"github.com/bazelbuild/bazel-gazelle/internal/language/proto"
 	"github.com/bazelbuild/bazel-gazelle/language"
+	"github.com/bazelbuild/bazel-gazelle/language/proto"
 	"github.com/bazelbuild/bazel-gazelle/rule"
 )
