+++ language/proto/gen/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["gen_known_imports.go"],
-    importpath = "github.com/bazelbuild/bazel-gazelle/internal/language/proto/gen",
+    importpath = "github.com/bazelbuild/bazel-gazelle/language/proto/gen",
     visibility = ["//visibility:private"],
     deps = ["//label:go_default_library"],
 )
