+++ language/proto/testdata/default_visibility/foo.proto
