+++ language/proto/testdata/multiple_packages/default_mode/BUILD.want
