+++ language/proto/testdata/multiple_packages/default_mode/foo.proto
