+++ language/proto/testdata/multiple_packages/package_mode/BUILD.want
