+++ language/proto/testdata/multiple_packages/package_mode/bar1.proto
