+++ language/proto/testdata/multiple_packages/package_mode/bar2.proto
