+++ language/proto/testdata/multiple_packages/package_mode/foo1.proto
