+++ language/proto/testdata/multiple_packages/package_mode/foo2.proto
