+++ language/proto/testdata/multiple_packages/package_mode_group/BUILD.want
