+++ language/proto/testdata/multiple_packages/package_mode_group/bar1.proto
