+++ language/proto/testdata/multiple_packages/package_mode_group/bar2.proto
