+++ language/proto/testdata/multiple_packages/package_mode_group/foo1.proto
