+++ language/proto/testdata/protos/BUILD.want
