+++ language/proto/testdata/protos/foo.proto
